% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreements_database.R
\docType{data}
\name{agreements}
\alias{agreements}
\title{Agreements Database Documentation}
\format{
The agreements database is a list that contains a shorter version
of the following 3 datasets: IEADB, GNEVAR, ECOLEX.
}
\usage{
agreements
}
\description{
Agreements Database Documentation
}
\details{
These datasets are to be used for testing and
running examples and vignettes in qData.
For the complete database and more datasets on environmental agreements
please refer to the \code{{qEnviron}} package.
\describe{
\item{IEADB: }{A dataset with 200 observations and the following 8 variables: IEADB_ID, Title, Beg, L, D, Signature, Force, qID.}
\item{GNEVAR: }{A dataset with 200 observations and the following 10 variables: GNEVAR_ID, Title, Beg, End, L, J, D, Signature, Force, qID.}
\item{ECOLEX: }{A dataset with 200 observations and the following 8 variables: ECOLEX_ID, Title, Beg, L, J, Signature, Force, qID.}
}
}
\keyword{datasets}
