% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambigdate.R
\name{ambig_date}
\alias{ambig_date}
\title{Create nested vectors of dates from vague date inputs}
\usage{
ambig_date(x)
}
\arguments{
\item{x}{String vector of potential dates.}
}
\value{
Nested vector of POSIXct dates that includes
}
\description{
Create nested vectors of dates from vague date inputs
}
\examples{
\dontrun{
checkdates <- function(){
testdates <- c("2010-01-01", "2010", "2010-01",
"2010-01-00", "2010-00-00",
"2010-01-??", "2010-??-??",
"2010:2011",
"2010-01-01_03", "2010-01_03",
"2010-01-01:03", "2010-01:03",
"9999-12-31", "2599-12-31",
"1712-01-01", "712-01-01", "712 AD",
"-1712-01-01", "-712-01-01", "712 BC")
datestest <- tibble::tibble(
origin = testdates,
lubridate = lubridate::as_date(testdates),
anytime = anytime::anydate(testdates),
qDatr = qDatr::ambig_date(testdates)
) datestest \%>\% print(n = 25)}
}
}
