% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fisher_info.R
\name{ApproxFI}
\alias{ApproxFI}
\title{Numerically approximate the fisher information matrix using finite difference method}
\usage{
ApproxFI(masks, theta, mm, dat, delta = 1e-04)
}
\arguments{
\item{masks}{A masks object from masks().}

\item{theta}{A theta vector.}

\item{mm}{A model matrix from causl::mm().}

\item{dat}{Input data to calculate log-likelihood on.}

\item{delta}{The infinitesimal delta used in finite difference method.}
}
\value{
The sum of log-likelihood.
}
\description{
Numerically approximate the fisher information matrix using finite difference method
}
\examples{


}
