% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{call}
\alias{call}
\title{Call 'many' packages, sources, and data}
\arguments{
\item{pkg}{A character vector of package name.
For multiple packages,
please declare package names as a vector (e.g. c("pkg1", "pkg2)).}

\item{database}{A database from one of the many packages.}

\item{dataset}{A dataset in a database from one of the many packages.
NULL by default.
That is, all datasets in the database are used.
For multiple datasets, please declare datasets as a vector
(e.g. c("dataset1", "dataset2)).}

\item{treaty_type}{The type of treaties to be returned.
NULL, by default.
Other options are "bilateral" or "multilateral".}

\item{develop}{Would you like to download the develop
version of the package?
FALSE by default.
If TRUE, the function downloads the develop version of package from GitHub.}

\item{variable}{Would you like to get one, or more, specific variables
present in one or more datasets in the 'many' database?
NULL by default.
For multiple variables, please declare variable names as a vector.}

\item{actor}{An actor variable in dataset.
NULL by default.
If declared, a tibble of the traties and their member actors is returned.}
}
\value{
The \code{call_} functions return tibbles with the respective information.
}
\description{
The \code{call_} functions in \code{{manydata}} allows users to call,
install, and update different 'many' packages, as well as additional
information on databases and datasets across 'many packages'.
}
\examples{
\donttest{
#call_packages()
#call_packages("manyenviron")
call_sources(emperors)
membs <- tibble::tibble( manyID = c("ROU-RUS[RFP]_1901A",
"ROU-RUS[RFP]_1901A", "GD16FI_1901A"),
StateID = c("ROU", "RUS", "DNK"),
Title = c("Convention Between Roumania And Russia Concerning Fishing
In The Danube And The Pruth",
"Convention Between Roumania And Russia Concerning Fishing
In The Danube And The Pruth",
"Convention Between The Governments Of Denmark And
The United Kingdom Of Great Britain
And Northern Ireland For Regulating The Fisheries
Of Their Respective Subjects Outside
Territorial Waters In The Ocean Surrounding The Faroe Islands"),
Beg = c("1901-02-22", "1901-02-22", "1901-06-24"))
call_treaties(membs)
call_treaties(membs, treaty_type = "bilaterals",
variable = c("Title", "Beg"))
call_treaties(membs, variable = c("Title", "Beg"), actor = "StateID")
}
}
