% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_packages.R
\name{call_releases}
\alias{call_releases}
\title{Call releases historical milestones/releases}
\source{
https://benalexkeen.com/creating-a-timeline-graphic-using-r-and-ggplot2/
}
\usage{
call_releases(repo, begin = NULL, end = NULL)
}
\arguments{
\item{repo}{the github repository to track, e.g. "globalgov/manydata"}

\item{begin}{When to begin tracking repository milestones.
By default NULL, two months before the first release.}

\item{end}{When to end tracking repository milestones.
By default NULL, two months after the latest release.}
}
\value{
A ggplot graph object
}
\description{
The function will take a data frame that details this information,
or more usefully, a Github repository listing.
}
\details{
The function creates a project timeline graphic using ggplot2
with historical milestones and milestone statuses gathered from a
specified GitHub repository.
}
\examples{
\donttest{
#call_releases("globalgov/manydata")
#call_releases("manypkgs")
}
}
\seealso{
Other call_:
\code{\link[=call_packages]{call_packages()}},
\code{\link[=call_treaties]{call_treaties()}}
}
\concept{call_}
