% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_sources.R
\name{call_sources}
\alias{call_sources}
\alias{call_citations}
\title{Call sources and citations}
\usage{
call_sources(x)

call_citations(x, output = c("console", "help"))
}
\arguments{
\item{x}{A datacube or dataset}

\item{output}{Whether the output should be formatted for "console"
or the "help" page.}
}
\description{
These functions call any source or citation information that is available
for a datacube or dataset.
The function can be used on its own to the console,
called during another function call such as \code{consolidate()} or \code{pluck()},
or is used to automatically and consistently populate help files.
}
