% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_agreements.R
\name{code_agreements}
\alias{code_agreements}
\title{Code Agreement Titles}
\usage{
code_agreements(title, date, dataset = NULL)
}
\arguments{
\item{title}{title column variable}

\item{date}{date column variable}

\item{dataset}{name of the dataset, optional}
}
\description{
Creates an ID column that contains information on the
parties to an agreement, the type of agreement, the date
and the linkage to other agreements in the dataset.
}
\examples{
IEADB <- dplyr::slice_sample(qEnviron::agreements$IEADB, n = 10)
IEADB <- code_agreements(IEADB$Title, IEADB$Signature, IEADB)
}
