% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_agreements.R
\name{code_areas}
\alias{code_areas}
\title{Code the Treaty Areas}
\usage{
code_areas(x)
}
\arguments{
\item{x}{A character vector of treaty title}
}
\value{
A character vector of the treaty area
}
\description{
Identify the areas the treaty title refers to.
}
\examples{
\dontrun{
IEADB$area <- Code_areas(IEADB$Title)
}
}
