% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_agreements.R
\name{code_dates}
\alias{code_dates}
\title{Creates Numerical IDs from Signature Dates}
\usage{
code_dates(title, date)
}
\arguments{
\item{title}{A title variable}

\item{date}{A date variable}
}
\value{
A character vector with condensed dates
}
\description{
Agreements should have a unique identification number that is meaningful,
we condense their signature dates to produce this number.
}
\examples{
IEADB <- dplyr::slice_sample(qEnviron::agreements$IEADB, n = 10)
IEADB$uID <- code_dates(IEADB$Title, IEADB$Signature)
}
