% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_agreements.R
\name{code_known_agreements}
\alias{code_known_agreements}
\title{Known agreements abbreviation}
\usage{
code_known_agreements(x)
}
\arguments{
\item{x}{A character vector of treaty title}
}
\value{
A character vector with abbreviation of known treaties
}
\description{
Some agreements have known abbreviations that facilitate their identification.
}
\examples{
\dontrun{
IEADB$abrevv <- code_known_agreements(IEADB$titles)
}
}
