% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_agreements.R
\name{code_parties}
\alias{code_parties}
\title{Code Agreement Parties}
\usage{
code_parties(title)
}
\arguments{
\item{title}{A character vector of treaty titles}
}
\value{
A character vector of parties that are mentioned in the treaty title
}
\description{
Identify the countries that are part of an agreement.
}
\examples{
IEADB <- dplyr::slice_sample(qEnviron::agreements$IEADB, n = 10)
IEADB$parties <- code_parties(IEADB$Title)
}
