% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_agreements.R
\name{code_topic}
\alias{code_topic}
\title{Code Agreement Topic}
\usage{
code_topic(x)
}
\arguments{
\item{x}{A character vector of treaty title}
}
\value{
A character vector of the treaty topic abbreviation.
}
\description{
Identify the main environmental issue the treaty is tackling.
}
\examples{
\dontrun{
IEADB$topic <- code_topic(IEADB$Title)
}
}
