% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_agreements.R
\name{code_type}
\alias{code_type}
\title{Code Agreement Type}
\usage{
code_type(x)
}
\arguments{
\item{x}{A character vector of treaty title}
}
\value{
A character vector of the treaty type
}
\description{
Identify the type of international agreement.
}
\examples{
IEADB <- dplyr::slice_sample(qEnviron::agreements$IEADB, n = 10)
IEADB$type <- code_type(IEADB$Title)
}
