% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_agreements.R
\name{codebook_agreements}
\alias{codebook_agreements}
\title{Codebook for qID}
\usage{
codebook_agreements()
}
\value{
a help document with the codebook for qID in the viewer pane
codebook_agreements()
}
\description{
Codebook for qID
}
