% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\alias{collapse_select}
\alias{collapse_full}
\alias{collapse_consensus}
\title{Collapse a database into a single data frame/tibble}
\usage{
collapse_select(dbase, dset)

collapse_full(dbase, key, resolve = NULL)

collapse_consensus(dbase, key, resolve = NULL)
}
\arguments{
\item{dbase}{A qPackage database object}

\item{dset}{A dataset label from within that database}

\item{key}{An ID column to collapse by}

\item{resolve}{How do you want differences to be resolved?
It takes "max", "min", "mean", "mode" and "median" options.}
}
\description{
Collapse a database into a single data frame/tibble
}
