% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_left}
\alias{collapse_left}
\title{Collapse a database into a single data frame/tibble}
\usage{
collapse_left(dbase, key, resolve = NULL)
}
\arguments{
\item{dbase}{A qPackage database object (can be subsetted)}

\item{key}{An ID column to collapse by}

\item{resolve}{How do you want differences to be resolved?
It takes "max", "min", "mean", "mode" and "median" options.}
}
\value{
A tibble with the collapsed data.
}
\description{
collapse_left() allows users to perform a left join on one or
more datasets in a database for further analysis.
}
\examples{
\dontrun{
collapse_left(qStates::states[1:2], "ID")
}
}
