% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_select}
\alias{collapse_select}
\title{Select one or multiple columns from a list of datasets}
\usage{
collapse_select(dbase, dset)
}
\arguments{
\item{dbase}{A qPackage database object}

\item{dset}{A dataset label from within that database}
}
\value{
A tibble with the collapsed data.
}
\description{
collapse_select allows users to select one or multiple columns
from a dataset in a database.
}
\examples{
\dontrun{
collapse_select(qStates::states, "COW")
}
}
