% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\alias{compare_data}
\alias{compare_overlap}
\alias{compare_missing}
\alias{compare_categories}
\title{Compare 'many' databases and datasets}
\usage{
compare_data(database, dataset = NULL)

compare_overlap(database, dataset = NULL, key = "manyID")

compare_missing(database, dataset = NULL, variable = "all")

compare_categories(
  database,
  dataset = NULL,
  key = "manyID",
  variable = "all",
  category = "all"
)
}
\arguments{
\item{database}{A database from one of the many packages.}

\item{dataset}{A dataset in a database from one of the many packages.
NULL by default.
That is, all datasets in the database are used.
A list of 2 or more datasets present in the database.}

\item{key}{A variable key to join datasets.
"manyID" by default.}

\item{variable}{Would you like to focus on one, or more, specific variables
present in one or more datasets in the 'many' database?
By default "all".
For multiple variables, please declare variable names as a vector.}

\item{category}{Would you like to focus on one specific code category?
By default "all" are returned.
Other options include "confirmed", "unique", "missing", "conflict",
or "majority".
For multiple variables, please declare categories as a vector.}
}
\value{
The compare functions return tibbles with the respective comparative
information, or the same information plotted appropriately.
}
\description{
The \code{compare_} functions in \code{{manydata}} allows users to quickly compare
different information on databases and/or datasets across 'many packages'.
These include comparison for:
\itemize{
\item data summaries
\item missing observations
\item categories
Each of these functions is accompanied by an appropriate printing method that
allows users to visualize the comparisons.
}
}
\details{
\code{compare_data()} compares the number of observations, variables,
the earliest and latest date in each dataset in a 'many' database.

\code{compare_overlap()} compares the overlap between "key" observations
in each dataset in a 'many' database.

\code{compare_missing()} compares the missing observations for variables
in each dataset in a 'many' database.

Confirmed values are the same in all datasets in database.
Unique values appear once in datasets in database.
Missing values are missing in all datasets in database.
Conflict values are different in the same number of datasets in database.
Majority values have the same value in multiple, but not all,
datasets in database.
}
\section{Functions}{
\itemize{
\item \code{compare_data()}: Compare data summaries for 'many' data

\item \code{compare_overlap()}: Compare the overlap between datasets in 'many' databases

\item \code{compare_missing()}: Compare missing observations for 'many' data

\item \code{compare_categories()}: Compare categories in 'many' databases

}}
\examples{
compare_data(emperors)
compare_overlap(emperors, key = "ID")
plot(compare_overlap(emperors, key = "ID"))
compare_missing(emperors)
plot(compare_missing(emperors))
compare_categories(emperors)
plot(compare_categories(emperors))
}
