% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_data}
\alias{compare_data}
\title{Compare data summaries for 'many' data}
\usage{
compare_data(database, dataset = NULL)
}
\arguments{
\item{database}{A database from one of the many packages.}

\item{dataset}{A dataset in a database from one of the many packages.
NULL by default.
That is, all datasets in the database are used.
A list of 2 or more datasets present in the database.}
}
\value{
\code{compare_data()} returns a tibble with information about each dataset
including the number of observations, the number of variables,
the earliest date, and the latest date in all observations.
}
\description{
Compare data summaries for 'many' data
}
\details{
\code{compare_data()} compares the number of observations, variables,
the earliest and latest date in each dataset in a 'many' database.
}
\examples{
\donttest{
compare_data(emperors)
}
}
\seealso{
Other compare_: 
\code{\link{compare_categories}()},
\code{\link{compare_missing}()},
\code{\link{compare_overlap}()}
}
\concept{compare_}
