% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat-helpers.R
\name{create_local_package}
\alias{create_local_package}
\title{Test Helpers for Testthat}
\source{
https://www.tidyverse.org/blog/2020/04/self-cleaning-test-fixtures/#usethis-and-create_local_package
}
\usage{
create_local_package(qTest = fs::file_temp(), env = parent.frame())
}
\arguments{
\item{qTest}{test package}

\item{env}{new temporary environment}
}
\value{
No objects are returned. Temporary package is delected after tests.
}
\description{
Interactive tests for developer function
}
