% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat-helpers.R
\name{create_test_package}
\alias{create_test_package}
\title{Test Helpers for Testthat}
\source{
https://www.tidyverse.org/blog/2020/04/self-cleaning-test-fixtures/#usethis-and-create_local_package
}
\usage{
create_test_package(
  qTest = fs::file_temp(pattern = "testpkg"),
  env = parent.frame(),
  rstudio = FALSE
)
}
\arguments{
\item{qTest}{test package}

\item{env}{new temporary environment}
}
\value{
No objects are returned. Temporary package is delected after tests.
}
\description{
Interactive tests for developer function
}
