% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_data.R
\name{data_source}
\alias{data_source}
\title{Displays names of the database/datasets and source material of qPackage data}
\usage{
data_source(pkg, database = NULL, dataset = NULL)
}
\arguments{
\item{pkg}{character string of the qPackage to report data on. Mandatory
input.}

\item{database}{vector of character strings of the qPackage to report data a
specific database in a qPackage. If Null, report_data returns a summary
of all databases in the qPackage. Null by default.}

\item{dataset}{character string of the qPackage to report data on a specific
dataset in a specific database of a qPackage. If Null and database is
specified, returns database level metadata. Null by default.}
}
\value{
A dataframe with the data sources.
}
\description{
To quickly report the names of the data structures in a package as well
as their source.
}
\examples{
data_source(pkg = "qStates", database = "states", dataset = "COW")
}
