% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_plot.R
\name{dbcomp}
\alias{dbcomp}
\title{Compare variable(s) in a 'many' database}
\usage{
dbcomp(database, variable, key = "manyID", category = "all")
}
\arguments{
\item{database}{A many database.}

\item{variable}{One, or more, variables present in multiple datasets
in a many database.
For multiple variables, please declare as a list
(i.e. c("variable1", "variable2"))}

\item{key}{A variable key to join datasets by, "manyID" by default.}

\item{category}{Would you like to focus on one specific code category?
By default "all" are returned.
Other options include "confirmed", "unique", "missing", "conflicting",
or "majority".}
}
\value{
A tibble with the varaible(s) profile.
}
\description{
Compare variable(s) in a 'many' database
}
\details{
The tibble returns the values for all observations
in the variable(s) selected from a 'many' package databasethat belong to all,
or a specific, code category.
Code categories include:
Confirmed values are the same in all datasets in database.
Unique values appear once in datasets in database.
Missing values are missing in all datasets in database.
Conflicting values are different in the same number of datasets in database.
majority values have the same value in multiple, but not all,
datasets in database.
}
\examples{
dbcomp(database = emperors, variable = "Beg", key = "ID")
dbcomp(database = emperors, variable = c("Beg", "End"), key = "ID",
category = "conflict")
#dbcomp(database = manyenviron::agreements, variable = "Title",
#category = "conflict")
}
