% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manydata-defunct.R
\name{defunct}
\alias{defunct}
\alias{coalesce_compatible}
\alias{network_map}
\alias{db_plot}
\alias{db_comp}
\alias{retrieve_bilaterals}
\alias{retrieve_multilaterals}
\alias{retrieve_links}
\alias{retrieve_membership_list}
\alias{data_source}
\alias{get_packages}
\alias{open_codebook}
\alias{data_contrast}
\alias{data_evolution}
\alias{plot_releases}
\alias{code_extend_glove}
\title{Functions that have been renamed, superseded, or are no longer working}
\usage{
coalesce_compatible(.data)

network_map(.data)

db_plot(.data)

db_comp(.data)

retrieve_bilaterals(.data)

retrieve_multilaterals(.data)

retrieve_links(.data)

retrieve_membership_list(.data)

data_source(.data)

get_packages(.data)

open_codebook(.data)

data_contrast(.data)

data_evolution(.data)

plot_releases(.data)

code_extend_glove(titles, var, req_f1 = 0.8, rarity_threshold = 8)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Generally these functions have been superseded or renamed.
Upon using them, a message is provided directing the user to the new function.
However, at this stage of package development,
we generally clear older defunct functions at each minor release,
and so you are strongly encouraged to use the new functions/names/syntax
wherever possible and update your scripts accordingly.
}
\section{Functions}{
\itemize{
\item \code{coalesce_compatible()}: Deprecated on 2022-10-18.

\item \code{network_map()}: Removed on 2023-06-15.

\item \code{db_plot()}: Removed on 2023-06-19.

\item \code{db_comp()}: Removed on 2023-06-19.

\item \code{retrieve_bilaterals()}: Removed on 2023-06-19.

\item \code{retrieve_multilaterals()}: Removed on 2023-06-19.

\item \code{retrieve_links()}: Removed on 2023-06-19.

\item \code{retrieve_membership_list()}: Removed on 2023-06-19.

\item \code{data_source()}: Removed on 2023-06-19.

\item \code{get_packages()}: Removed on 2023-06-19.

\item \code{open_codebook()}: Removed on 2023-06-19.

\item \code{data_contrast()}: Removed on 2023-06-19.

\item \code{data_evolution()}: Removed on 2023-06-19.

\item \code{plot_releases()}: Renamed on 2023-07-03.

\item \code{code_extend_glove()}: Removed on 2026-10-02.

}}
\keyword{internal}
