% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\name{describe}
\alias{describe}
\alias{mreport}
\alias{describe_datacube}
\title{Data reports for datacubes and datasets with 'mdate' variables}
\usage{
mreport(data)

describe_datacube(datacube)
}
\arguments{
\item{data}{A \code{{tibble}} or a \code{{data.frame}}.}

\item{datacube}{A datacube}
}
\value{
A data report of class 'mreport'.
}
\description{
These functions provide meta level descriptions of datacubes or datasets.
\code{mreport()} creates a properly formatted data report for datasets which contain 'mdate'
class objects, alongside other object classes.
\code{describe_datacube()} prints a text description of the datasets in a
datacube.
}
\details{
'mreport' displays the variable's name,
the variable type, the number of observations per variable,
the number of missing observations for variable,
and the percentage of missing observations in variable.
}
\examples{
mreport(emperors)
}
