% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{export_data}
\alias{export_data}
\title{Save a cleaned data object in the new q package}
\usage{
export_data(..., overwrite = FALSE, compress = "bzip2")
}
\arguments{
\item{...}{Unquoted names of existing objects to save}

\item{overwrite}{Whether to overwrite any existing objects saved}

\item{compress}{Compression formula}
}
\description{
Save a cleaned data object, consistent with the qDatr ecosystem, ready to be lazy-loaded
and create scripts for documenting and testing that object within the new q package
}
\examples{
\dontrun{
TODO
}
}
