% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extraction}
\alias{extraction}
\alias{extract_bilaterals}
\alias{extract_multilaterals}
\title{Extracting international agreements}
\usage{
extract_bilaterals(membs)

extract_multilaterals(membs)
}
\arguments{
\item{membs}{A memberships dataset from one of the many packages}
}
\value{
A tibble of bilateral agreements

A tibble of multilateral agreements
}
\description{
This set of functions help extracting bilateral agreements or
multilateral agreements from a membership database
in the many packages universe.
}
\details{
The function helps extracting bilateral agreements
from a membership database in the many packages universe.

The function helps extracting multilateral agreements
from a membership database in the many packages universe.
}
\examples{
membs <- tibble::tibble(CountryID = c("ROU", "RUS", "DNK"),
manyID = c("ROU-RUS[RFP]_1901A", "ROU-RUS[RFP]_1901A", "GD16FI_1901A"),
Title = c("Convention Between Roumania And Russia Concerning Fishing
In The Danube And The Pruth",
"Convention Between Roumania And Russia Concerning Fishing
In The Danube And The Pruth",
"Convention Between The Governments Of Denmark And
The United Kingdom Of Great Britain
And Northern Ireland For Regulating The Fisheries
Of Their Respective Subjects Outside
Territorial Waters In The Ocean Surrounding The Faroe Islands"),
Beg = c("1901-02-22", "1901-02-22", "1901-06-24"),
End = c(NA, NA, NA))
extract_bilaterals(membs)
membs <- tibble::tibble(CountryID = c("ROU", "RUS", "DNK"),
manyID = c("ROU-RUS[RFP]_1901A", "ROU-RUS[RFP]_1901A", "GD16FI_1901A"),
Title = c("Convention Between Roumania And Russia Concerning Fishing
In The Danube And The Pruth",
"Convention Between Roumania And Russia Concerning Fishing
In The Danube And The Pruth",
"Convention Between The Governments Of Denmark And
The United Kingdom Of Great Britain
And Northern Ireland For Regulating The Fisheries
Of Their Respective Subjects Outside
Territorial Waters In The Ocean Surrounding The Faroe Islands"),
Beg = c("1901-02-22", "1901-02-22", "1901-06-24"),
End = c(NA, NA, NA))
extract_multilaterals(membs)
}
