% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate.R
\name{favour}
\alias{favour}
\alias{favor}
\title{Favour datasets}
\usage{
favour(database, x)

favor(database, x)
}
\arguments{
\item{database}{A many database}

\item{x}{Would you like a dataset to be favoured over others?
In this case the dataset selected becomes the reference for
the first non NA value.
If more than one dataset is declared,
please add datasets increasing order of importance
(.i.e. last dataset should be favoured over previous).}
}
\value{
The same database with datasets re-ordered accordingly
}
\description{
Favour datasets
}
\examples{
favour(emperors, "UNRV")
favour(emperors, c("wikipedia", "UNRV", "britannica"))
}
