% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{filter_datacube}
\alias{filter_datacube}
\title{Filtering datacube datasets to a certain date}
\usage{
filter_datacube(datacube, date = Sys.Date())
}
\arguments{
\item{datacube}{A datacube,
i.e. a list of data frames with Begin and End date variables.}

\item{date}{A date (of class Date or character) at which to filter the datacube.}
}
\description{
Filtering datacube datasets to a certain date
}
\examples{
filter_datacube(emperors, date = "0100")
}
