% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_find.R
\name{find}
\alias{find}
\alias{find_ID}
\alias{find_common_ID}
\alias{find_duplicates}
\title{Find elements within manydata}
\usage{
find_ID(df, id_col = "ID")

find_common_ID(..., id_col = "ID")

find_duplicates(df, id_col = "ID")
}
\arguments{
\item{df}{A data frame to be scored.}

\item{id_col}{The name of the column containing IDs.
Default is "ID".}

\item{...}{Data frames to compare}
}
\description{
Find elements within manydata
}
\examples{
find_duplicates(emperors$Wikipedia)
}
