% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_find.R
\name{find_year}
\alias{find_year}
\title{Creates Numerical IDs from Signature Dates}
\usage{
find_year(date)
}
\arguments{
\item{date}{A date variable}
}
\value{
A character vector with condensed dates
}
\description{
Agreements should have a unique identification
number that is meaningful, we condense their
signature dates to produce this number.
}
\examples{
\dontrun{
IEADB <- dplyr::slice_sample(manyenviron::agreements$IEADB, n = 10)
code_dates(IEADB$Title)
}
}
