% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{import_data}
\alias{import_data}
\title{Imports and establishes preparation of raw data}
\usage{
import_data(
  name = "DATASET",
  path = NULL,
  delete_original = FALSE,
  open = rlang::is_interactive()
)
}
\arguments{
\item{name}{Intended (short)name of the dataset}

\item{path}{Path to raw data file. If left unspecified, a dialog box is raised to select the file via the system.}

\item{open}{Whether the resulting preparation script will be opened}
}
\value{
This function...
}
\description{
Create a data-raw folder and provide templates that make easier for setting up the data cleaning
and wrangling, consistent with the qDatr ecosystem
}
\details{
The function...
}
\examples{
\dontrun{
TODO
}
}
