% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{interleave}
\alias{interleave}
\title{Interleaving two vectors by position}
\usage{
interleave(vect, pos, elems = NA)
}
\arguments{
\item{vect}{Main vector}

\item{pos}{Positions to be inserted}

\item{elems}{Elements to be inserted at those positions.
By default, these are NAs (missing values).}
}
\value{
A vector the length of the sum of \code{vect}
and \code{pos}.
}
\description{
Resets the century of (unlikely) future events
}
\examples{
interleave(1:5, c(2,4))
}
