% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_agreements.R
\name{known_agreement_dates}
\alias{known_agreement_dates}
\title{Code Known Agreement Signature Date}
\usage{
known_agreement_dates(x)
}
\arguments{
\item{x}{A character vector of treaty title}
}
\value{
A character vector of treaty signature dates
}
\description{
Identify known agreements and assign the signature date
}
\examples{
\dontrun{
IEADB$abrev_dates <- known_agreements_dates(IEADB$titles)
}
}
