% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_agreements.R
\name{known_agreements}
\alias{known_agreements}
\title{Known agreements abbreviation}
\usage{
known_agreements(x)
}
\arguments{
\item{x}{A character vector of treaty title}
}
\value{
A character vector with abbreviation of known treaties
}
\description{
Known agreements abbreviation
}
