% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_calculate_fisher_info.R
\name{ll_sum}
\alias{ll_sum}
\title{Calculate the sum of log-likelihood}
\usage{
ll_sum(masks, theta, mm, dat)
}
\arguments{
\item{masks}{A masks object from masks().}

\item{theta}{A theta vector.}

\item{mm}{A model matrix from causl::mm().}

\item{dat}{Input data to calculate log-likelihood on.}
}
\value{
The sum of log-likelihood.
}
\description{
Calculate the sum of log-likelihood
}
\examples{


}
