% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{new_author_orcid}
\alias{new_author_orcid}
\title{Helper function for adding an author to the current package by using the
ORCID number as an input.}
\usage{
new_author_orcid(ORCID, role = NULL, email = NULL)
}
\arguments{
\item{ORCID}{Character string of the ORCID number of the author to add}

\item{role}{Character vector of role(s) the author has in the project}

\item{email}{Character strin of the author's email. ORCID is preferred.}
}
\value{
Adds a new author to the description file of the package
}
\description{
Helper function for adding an author to the current package with the ORCID
number of the author.
}
\details{
The function adds an author to the description file of the current
package based on information gathered from ORCID.
}
