% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_agreements.R
\name{order_agreements}
\alias{order_agreements}
\title{Extracts Orders from Agreements}
\usage{
order_agreements(title)
}
\arguments{
\item{title}{A character vector of treaty title}
}
\value{
A character vector with meangniful numbers from titles
}
\description{
Identifies and extracts meaningful numbers from agreements titles that
are important information about the order of the agreement.
}
