% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_}
\alias{plot_}
\alias{plot_agreements}
\alias{plot_memberships}
\alias{plot_lineage}
\alias{plot_map}
\title{Plot agreements network}
\usage{
plot_agreements(dataset, treaty_type = NULL, key = "manyID", layout = "circle")

plot_memberships(
  dataset,
  actor = "stateID",
  treaty_type = NULL,
  key = "manyID",
  layout = "bipartite"
)

plot_lineage(dataset, treaty_type = NULL, key = "manyID", layout = "nicely")

plot_map(
  dataset,
  actor = "stateID",
  treaty_type = NULL,
  date = "2019-12-31",
  theme = "light"
)
}
\arguments{
\item{dataset}{A dataset from one of the many packages
or a "consolidated" database.}

\item{treaty_type}{The type of treaties to be returned.
NULL, by default.
Other options are "bilateral" or "multilateral".}

\item{key}{An ID column to collapse by.
By default "manyID".}

\item{layout}{How do you want the plot to look like?
An \code{{ggraph}} layout algorithm.
If not declared, reasonable defaults are used.}

\item{actor}{An actor variable.
"stateID", by default.}

\item{date}{String date from the network snapshot.
Used by \code{{cshapes}} to plot the correct map.
By default, 2019-12-31.
Date can be between 1886-01-01 and 2019-12-31.}

\item{theme}{Theme you would like to use to plot the graph.
bey defalt, "light".
Available themes are "light", "dark", and "earth".}
}
\value{
A network of agreements' relations.

A network of agreements' memberships.

A plot of agreements' lineages.

A map of a country level geographical network.
}
\description{
Facilitates plotting of 'many' data.
}
\details{
\code{plot_map()} creates a plot of the a unimodal geographical network
at a single point in time.
}
\examples{
\donttest{
#agreements <- dplyr::filter(manyenviron::agreements$ECOLEX,
#Begin > "2000-01-01" & Begin < "2000-12-12")
#plot_agreements(agreements)
}
\donttest{
#memberships <- dplyr::filter(manyenviron::memberships$ECOLEX_MEM,
#Begin > "2000-01-01" & Begin < "2000-01-31")
#plot_memberships(memberships)
}
\donttest{
#lineage <- dplyr::filter(manyenviron::agreements$HUGGO,
#Begin > "2000-01-01", Begin < "2001-12-31")
#plot_lineage(lineage)
}
\donttest{
#memberships <- dplyr::filter(manyenviron::memberships$ECOLEX_MEM,
#Begin > "2000-01-01" & Begin < "2000-12-12")
#plot_map(memberships, actor = "stateID") +
#ggplot2::labs(title = "Bilateral International Environmental Treaties Signed in the year 2000",
#subtitle = "Ecolex data")
}
}
