% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{qpackage_create}
\alias{qpackage_create}
\title{Create a new package in the qDatr ecosystem}
\usage{
qpackage_create(packageName, packageAuthor)
}
\arguments{
\item{packageName}{A string giving the desired name of the package, must start with "q"}

\item{packageAuthor}{A string, list or vector giving the package author(s), required}
}
\value{
A new package structure
}
\description{
Creates a new package in and consistent with the qDatr ecosystem
}
\details{
The function establishes many of the required files and folder structures
required for a qDatr-consistent data package.
}
\examples{
\dontrun{
qpackage_create("qStates", c("James Hollway", "Henrique Sposito"))
}
}
