% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{qtemplate}
\alias{qtemplate}
\title{Helper function for finding and rendering templates}
\usage{
qtemplate(
  template,
  save_as = template,
  data = list(),
  ignore = FALSE,
  path,
  open = rlang::is_interactive(),
  package = "qData"
)
}
\arguments{
\item{template}{Template called}

\item{save_as}{Path to where the rendered template should be saved}

\item{data}{Any elements to be entered into the template via Whisker}

\item{ignore}{For use with usethis::use_build_ignore()}

\item{path}{Path to where template was gathered from}

\item{open}{Whether the resulting template will be opened}

\item{package}{Package called}
}
\value{
A rendered template, saved into the correct folder
}
\description{
Helper function for finding and rendering templates from the qData package
}
\details{
This function is an adaptation of the usethis variant
for use in the qData ecosystem.
}
\examples{
\dontrun{
qtemplate("test_states.R",
save_as = fs::path("tests", "testthat", paste0("test_", dataset_name, ".R")),
open = FALSE,
path = getwd())
}
}
