% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate.R
\name{r_coalesce}
\alias{r_coalesce}
\title{Resolve Coalesce}
\usage{
r_coalesce(other_variables, out, key)
}
\arguments{
\item{other_variables}{A list of variables to be resolved}

\item{out}{A dataframe}

\item{key}{The ID column to collapse by. By default "qID_ref"}
}
\value{
The resolved dataframed or variable
}
\description{
Helper function for resolving a database into a dataframe with coalesce.
"coalesce" takes the first non-NA value
}
