% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate.R
\name{r_max}
\alias{r_max}
\title{Resolve Maximum}
\usage{
r_max(other_variables, out, key)
}
\arguments{
\item{other_variables}{A list of variables to be resolved}

\item{out}{A dataframe}

\item{key}{The ID column to collapse by. By default "treaty_ID"}
}
\value{
The resolved dataframed or variable
}
\description{
Helper function for resolving a database into a dataframe.
"maximum" takes the maximum non-NA value
}
