% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate.R
\name{r_median}
\alias{r_median}
\title{Resolve Median}
\usage{
r_median(other_variables, out, key)
}
\arguments{
\item{other_variables}{A list of variables to be resolved}

\item{out}{A dataframe}

\item{key}{The ID column to collapse by. By default "qID"}
}
\value{
The resolved dataframed or variable
}
\description{
Helper function for resolving a database into a dataframe.
"Median" takes the median non NA value
}
