% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{rearrange}
\alias{rearrange}
\title{Moving variables relative to others}
\usage{
rearrange(data, tomove, where = "last", ba = NULL)
}
\arguments{
\item{data}{First variable to be used, required.}

\item{tomove}{Variable(s) to be moved}

\item{where}{String that dictates position in relation to
reference variable. Can be one of: "last", "first", "before", or "after".}

\item{ba}{Optional string identifying reference variable
By default this is the system date, but can be specified.}
}
\value{
The data frame given by 'data' with the variables repositioned
}
\description{
Moves variables (columns) of a data frame to positions
relative to other variables in the data frame.
}
\examples{
\dontrun{
gneva.treat <- rearrange(gneva.treat, "L", "after", "X")
gneva.treat <- rearrange(gneva.treat, c("Cites","Amends","Supersedes"), "before", "Amended.by")
}
}
