% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_dates.R
\name{recent}
\alias{recent}
\title{#' Identify and complete date ranges
#'
#' @param d dates
#' @return a range of dates
ranged <- function(d) {

date_range <- function(start, finish){
lubridate::as_date(lubridate::as_date(start):lubridate::as_date(finish))
}

if(stringr::str_detect(d, "^[:digit:]{4}:[:digit:]{4}$")){ # year range
brackets <- stringr::str_split(d, ":")
start <- paste0(brackets[\link{1}]\link{1}, "-01-01")
finish <- paste0(brackets[\link{1}][2], "-12-31")
d <- date_range(start, finish)
} else if(stringr::str_detect(d, "^[:digit:]{4}-[:digit:]{2}:[:digit:]{2}$")){ # month range
brackets <- stringr::str_split(d, ":")
start <- paste0(brackets[\link{1}]\link{1}, "-01")
finish <- paste(stringr::str_split(start, "-")[\link{1}]\link{1},
brackets[\link{1}][2],
lubridate::days_in_month(as.numeric(brackets[\link{1}][2])),
sep = "-")
d <- date_range(start, finish)
} else if(stringr::str_detect(d, "^[:digit:]{4}-[:digit:]{2}-[:digit:]{2}:[:digit:]{2}$")){ # day range
brackets <- stringr::str_split(d, ":")
start <- brackets[\link{1}]\link{1}
finish <- paste(stringr::str_split(start, "-")[\link{1}]\link{1},
stringr::str_split(start, "-")[\link{1}][2],
brackets[\link{1}][2],
sep = "-")
d <- date_range(start, finish)
} else {
d
}
}
Resetting century of future events}
\usage{
recent(dates, sep = NULL)
}
\arguments{
\item{dates}{First variable to be used, required.
Can be a character or date variable.}

\item{sep}{Year from which to make a cut-off and return the last century.
By default this is the system date, but can be specified.}
}
\value{
A vector the same length as \code{dates}
}
\description{
Resets the century of (unlikely) future events
}
\details{
This function will use the system date as the cut-off
for identifying past or future events.
It is geared towards numeric dates (i.e. "12") not named dates (i.e. "Dec"),
though these dates should be correctly parsed regardless of order or separator.
It will return any datestamps of "9999-12-31",
which may be interpreted as unknown/future date, as is.
Please note that different types of separators in the same vector,
for example c("2004-12-12","12/12/2004", "12.12.2004"),
may confuse the algorithm.
In other words, this function is built for internally consistent codes.
}
\examples{
\dontrun{
recent(head(tfd_agree$Sign))
}
}
