% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{recollect}
\alias{recollect}
\title{Pastes unique string vectors}
\usage{
recollect(x, collapse = "_")
}
\arguments{
\item{x}{A vector}
}
\value{
A single value
}
\description{
For use with dplyr::summarise, for example
}
\details{
This function operates similarly to reunite,
but instead of operating on columns/observations,
it pastes together unique rows/observations.
}
\examples{
\dontrun{
}
}
