% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_agreements.R
\name{remove_dates}
\alias{remove_dates}
\title{Remove dates from agreement titles}
\usage{
remove_dates(title)
}
\arguments{
\item{title}{A character vector of treaty title}
}
\value{
A character vector without dates
}
\description{
Dates can cause an issue when trying to extract meaningful
numbers from titles for ordering purposes, this function removes dates.
}
