% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling_data.R
\name{repaint}
\alias{repaint}
\title{Fills missing data by lookup}
\usage{
repaint(df, id, var)
}
\arguments{
\item{df}{a dataframe}

\item{id}{a string identifying a column in the dataframe for indexing}

\item{var}{a string identifying a column or columns in the dataframe
to be filled}
}
\value{
A dataframe
}
\description{
Fills missing data where known by other observations with the same id/index
}
\examples{
\donttest{
data <- data.frame(ID = c(1,2,3,3,2,1),
                    One = c(1,NA,3,NA,2,NA),
                    Two = c(NA,"B",NA,"C",NA,"A"))
repaint(data, "ID", c("One","Two"))
}
}
