% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_data.R
\name{report_data}
\alias{report_data}
\title{Reports data within a qPackage}
\usage{
report_data(pkg)
}
\arguments{
\item{pkg}{character string of the qPackage to report data on}
}
\value{
a list with information on qPackage
}
\description{
Reports data within a qPackage
}
\details{
The function quickly report on the databases and datasets within a qPackage
}
\examples{
 
\dontrun{
report_data("qStates")
}
}
