% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{resolve}
\alias{resolve}
\alias{resolve_min}
\alias{resolve_max}
\alias{resolve_mean}
\alias{resolve_median}
\alias{resolve_mode}
\title{Resolve ranged dates into single vectors}
\usage{
resolve_min(var)

resolve_max(var)

resolve_mean(var)

resolve_median(var)

resolve_mode(var)
}
\arguments{
\item{dbase}{A qPackage database object}

\item{dset}{A dataset label from within that database}

\item{key}{An ID column to collapse by}
}
\description{
Resolve ranged dates into single vectors
}
