% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate.R
\name{resolve_random}
\alias{resolve_random}
\title{Resolve Random}
\usage{
resolve_random(other_variables, out, key)
}
\arguments{
\item{other_variables}{A list of variables to be resolved}

\item{out}{A dataframe}

\item{key}{The ID column to collapse by. By default "many_ID"}
}
\value{
The resolved dataframed or variable
}
\description{
Helper function for resolving a database into a dataframe.
"random" takes a random non NA value from sample
}
