% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{reunite}
\alias{reunite}
\title{Pastes unique string vectors}
\usage{
reunite(..., sep = "_")
}
\arguments{
\item{...}{Variables to pass to the function,
currently only two at a time}

\item{sep}{Separator when vectors reunited, by default "_"}
}
\value{
A single vector with unique non-missing information
}
\description{
A vectorised function for use with dplyr's mutate, etc
}
\examples{
\dontrun{
data <- data.frame(fir=c(NA, "two", "three", NA),
  sec=c("one", NA, "three", NA), stringsAsFactors = F)
transmutate(data, single = reunite(fir, sec))
}
}
