% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_samples.R
\name{run_MH_MCMC}
\alias{run_MH_MCMC}
\title{Simulate posterior samples using MCMC}
\usage{
run_MH_MCMC(
  startval,
  mcmc_pars,
  dat_obs,
  dat_exp,
  msks_obs,
  theta_obs,
  mm_obs,
  msks_exp,
  theta_exp,
  mm_exp,
  p_mu = rep(0, length(startval)),
  p_sig = diag(length(startval)),
  family,
  link,
  eta,
  prop_cov
)
}
\arguments{
\item{startval}{Initial values.}

\item{dat_obs}{Observational dataset.}

\item{dat_exp}{Experimental dataset.}

\item{msks_obs, msks_exp}{masks for parameter vectors with observational and experimental distributions}

\item{theta_obs, theta_exp}{masks for parameter vectors with observational and experimental distributions}

\item{p_mu, p_sigma}{vector and matrix for Gaussian proposal distribution}

\item{eta}{Learning rate on the observational data.}

\item{prop_cov}{optional covariance for proposal}

\item{n_iter}{Number of MCMC iterations.}
}
\value{
A table of posterior samples.
}
\description{
Simulate posterior samples using MCMC
}
\details{
Proposal is currently an arbitrary multiple of univariate second
derivatives (10).
}
\examples{
theta_sim_eta05 <- run_MH_MCMC(c(1,0.5), 1000, dat_obs = dat_obs, dat_exp = dat_exp, eta = 0.5)
}
