% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_setup.R
\name{setup_package}
\alias{setup_package}
\title{Create a new package in the qData ecosystem}
\usage{
setup_package(
  packageName = NULL,
  packageAuthorName = NULL,
  packageAuthorSurname = NULL,
  packageAuthorInstitution = NULL,
  ORCID = NULL,
  update = TRUE,
  path = getwd()
)
}
\arguments{
\item{packageName}{A string giving the desired name of the package,
must start with "q"}

\item{packageAuthorName}{A vector giving the package
author(s)' name(s)}

\item{packageAuthorSurname}{A vector giving the package
author(s)' surname(s)}

\item{packageAuthorInstitution}{A vector giving the package
author(s)' institution (only one per author)}

\item{ORCID}{A vector of strings of all the ORCID numbers of the authors.
Needs \code{{rorcid}} package to be installed. Takes precedence over manual
entries if specified.}

\item{update}{A logical indicating whether existing files should be
overwritten, by default TRUE.}

\item{path}{A string, if missing default is path to the working directory}
}
\value{
A new package structure
}
\description{
Creates a new package in, and consistent with, the qData ecosystem
}
\details{
The function establishes many of the required files and
folder structures required for a qData-consistent data package.
}
\examples{
\dontrun{
setup_package("qStates", "James Hollway, Henrique Sposito")
}
}
