% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_dates.R
\name{standardise_dates}
\alias{standardise_dates}
\title{Create nested vectors of dates from vague date inputs}
\usage{
standardise_dates(...)
}
\arguments{
\item{...}{One (ymd) or three (yyyy, mm, dd) variables}
}
\value{
Nested vector of POSIXct dates that includes a range of dates
}
\description{
Create nested vectors of dates for vague date inputs, ambiguous and ranged dates,
into a range of dates
}
\details{
The function seeks to convert ambiguous and ranged dates into a
range of dates, and extends the date parsing of other packages to more
historical and future dates.
The function allows only for dmy or ymd date formats at present, since mdy
may introduce errors.
}
\examples{
dates_comparison <- tibble::tribble(~Example, ~OriginalDate,
"A normal date", "2010-01-01",
"A historical date", "1712-01-01",
"A really historical date", "712-01-01",
"A very historical date", "-712-01-01",
"A clearly future date", "9999-12-31",
"A not so clearly future date", "2599-12-31")
dates_comparison \%>\% dplyr::mutate(
lubridate = lubridate::as_date(OriginalDate),
anytime = anytime::anydate(OriginalDate),
qData = qData::standardise_dates(OriginalDate)
) \%>\% print(n = 25)
}
