% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_titles.R
\name{standardise_titles}
\alias{standardise_titles}
\title{Standardise titles}
\usage{
standardise_titles(s, strict = FALSE, api_key = NULL)
}
\arguments{
\item{s}{A string}

\item{strict}{By default FALSE}

\item{api_key}{If google API key is provided, the function will
translate and return strings in english using google translator.}
}
\value{
A capitalised, trimmed and standardised string
}
\description{
Standardises words in a character title variable to improve readability,
facilitate string matching and enable more accurate comparisons
for variables in different datatsets.
}
\details{
The function capitalises words in the strings passed to it.
It trims white spaces from the start, middle and end of the strings.
Removes ambiguous punctions and symbols from strings.
All the strings are transformed into to ASCII character encoding.
Written numbers in ordinal form are transformed into numerical form.
If a google API key is provided as an argument, the function detects
strings in other languages and translates them to English.
}
\examples{
e <- standardise_titles("A treaty concerning things")
e==c("A Treaty Concerning Things")
}
