% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states_database.R
\docType{data}
\name{states}
\alias{states}
\title{states database documentation}
\format{
The states database is a list that contains the
following 3 datasets: GW, ISD, COW.
\describe{
\item{GW: }{A dataset with 150 observations and the following 5 variables: COW_Nr, ID, Beg, End, Label.}
\item{ISD: }{A dataset with 150 observations and the following 7 variables: COW_Nr, ID, Beg, End, Label, Micro, New.State.}
\item{COW: }{A dataset with 150 observations and the following 5 variables: COW_Nr, ID, Beg, End, Label.}
}
}
\usage{
states
}
\description{
states database documentation
}
\keyword{datasets}
