% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{use_qData}
\alias{use_qData}
\title{Createa a data file the new package for the qDatr ecosystem}
\usage{
use_qData(...)
}
\arguments{
\item{...}{Unquoted names of existing objects to save.}
}
\description{
Creates a data file with templates that make it consistent with the qDatr ecosystem
}
\examples{
\dontrun{
TODO
}
}
