% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{use_qData_raw}
\alias{use_qData_raw}
\title{Creates a dataraw file the new q package}
\usage{
use_qData_raw(name = "DATASET", open = rlang::is_interactive())
}
\arguments{
\item{name}{Intended (short)name of the dataset}

\item{open}{Whether the resulting preparation script will be opened}
}
\value{
A dataraw folder
}
\description{
Creates a dataraw file and provides templates that make it consistent with the qDatr ecosystem
}
\details{
The function loads raw data into a q package
}
\examples{
\dontrun{
TODO
}
}
