% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreements_plot.R
\name{agreements_plot}
\alias{agreements_plot}
\alias{plot_agreements}
\alias{plot_memberships}
\title{Plot agreements network}
\usage{
plot_agreements(dataset, treaty_type = NULL, layout = "concentric")

plot_memberships(
  dataset,
  actor = "StateID",
  treaty_type = NULL,
  layout = "concentric"
)
}
\arguments{
\item{dataset}{A dataset from one of the many packages
or a "consolidated" database.
Other options are "bilateral" or "multilateral".}

\item{treaty_type}{The type of treaties to be returned.
NULL, by default.
Other options are "bilateral" or "multilateral".}

\item{layout}{How do you want the plot to look like?
An igraph layout algorithm, currently defaults to 'concentric'.
Some other options are 'stress', 'bipartite', and 'alluvial'.
For more information please check ´?migraph::autographr´.}

\item{actor}{An actor variable.
"StateID", by default.}

\item{date}{String date from the network snapshot.
Used by \code{{cshapes}} to plot the correct map.
By default, 2019-12-31.
Date can be between 1886-01-01 and 2019-12-31.}

\item{theme}{Theme you would like to use to plot the graph.
bey defalt, "light".
Available themes are "light", "dark", and "earth".}
}
\value{
A network of agreements' relations.

A network of agreements' relations.
}
\description{
Facilitates plotting of 'many' data.
}
