% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_calculate_fisher_info.R
\name{ApproxFI}
\alias{ApproxFI}
\title{Approximate the fisher information matrix using finite differencing}
\usage{
ApproxFI(masks, theta, mm, dat, delta = 1e-04)
}
\arguments{
\item{masks}{A masks object from masks().}

\item{theta}{A theta vector.}

\item{mm}{A model matrix from causl::mm().}

\item{dat}{Input data to calculate loglikelihood on.}
}
\value{
The sum of log-likelihood.
}
\description{
Approximate the fisher information matrix using finite differencing
}
\examples{


}
