% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{call_packages}
\alias{call_packages}
\title{Call, download, and update many packages}
\usage{
call_packages(package, develop = FALSE)
}
\arguments{
\item{package}{A character vector of package name.
For multiple packages,
please declare package names as a vector (e.g. c("package1", "package2")).}

\item{develop}{Would you like to download the develop
version of the package?
FALSE by default.}
}
\value{
\code{call_packages()} returns a tibble with the 'many packages'
currently available.
If one or more package names are provided,
these will be installed from Github.
}
\description{
Call, download, and update many packages
}
\details{
\code{call_packages()} finds and download other packages
that belong to the many universe of packages.
It allows users to rapidly access the names and other
descriptive information of these packages.
If users intend to download and install a package listed,
they can type the package name within the function.
}
\examples{
\donttest{
#call_packages()
#call_packages("manyenviron")
}
}
\seealso{
Other call_: 
\code{\link{call_sources}()},
\code{\link{call_treaties}()}
}
\concept{call_}
