% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_agreements.R
\name{code_agreements}
\alias{code_agreements}
\title{Code agreements title}
\usage{
code_agreements(title, date, dataset = NULL)
}
\arguments{
\item{title}{title column variable}

\item{date}{date column variable}

\item{dataset}{name of the dataset}
}
\description{
Creates an ID column that contains information on the
parties, the topic, the date and the relations to other
agreements in the dataset.
}
