% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_agreements.R
\name{code_dates}
\alias{code_dates}
\title{Creates Unique ID numbers from dates}
\usage{
code_dates(x, date)
}
\arguments{
\item{x}{A title variable}

\item{date}{A date variable}
}
\value{
A character vector with condensed dates
}
\description{
Agreements should have a unique identification number that is meaningful,
we condense their signature date to produce this number.
}
\examples{
\dontrun{
IEADB$uID <- code_dates(IEADB$Title, IEADB$dates)
}
}
