% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_agreements.R
\name{code_dates}
\alias{code_dates}
\title{Cretes Unique ID numbers from dates}
\usage{
code_dates(x)
}
\arguments{
\item{x}{A date variable}
}
\value{
A character vector with condensed dates
}
\description{
Agreements should have a unique identification number that is meaningful,
so we condense their signature date to produce this number.
}
