% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_agreements.R
\name{code_linkage}
\alias{code_linkage}
\title{Code Agreement Linkages}
\usage{
code_linkage(title, date)
}
\arguments{
\item{title}{A character vector of treaty title}

\item{date}{A date variable}
}
\value{
A character vector of the agreements that are linked
}
\description{
Identify the linkage between amendments and protocols to a main agreement.
}
\details{
The function identifies duplicates via excluding a few
'predictable' words from strings, this maintains key words that are
then used to identify and link duplicates.
This is a choice that considers errors should lie on the side of false
negatives rather than false positives.
}
\examples{
IEADB <- dplyr::slice_sample(qEnviron::agreements$IEADB, n = 10)
code_linkage(IEADB$Title, IEADB$Signature)
}
