% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_agreements.R
\name{code_linkage}
\alias{code_linkage}
\title{Code Agreement Lineage}
\usage{
code_linkage(x, date)
}
\arguments{
\item{x}{A character vector of treaty title}

\item{date}{A date variable}
}
\description{
Identify the linkage between amendments and protocols to a main agreement.
}
\examples{
\dontrun{
IEADB$line <- code_linkage(IEADB$Title)
}
}
