% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_agreements.R
\name{code_linkage}
\alias{code_linkage}
\title{Code Agreement Lineage}
\usage{
code_linkage(x, date)
}
\arguments{
\item{x}{A character vector of treaty title}

\item{date}{A date variable}
}
\value{
A character vector of the agreements that are connected to one another
}
\description{
Identify the linkage between amendments and protocols to a main agreement.
}
\details{
The function identifies duplicates via excluding a few 'predictable' words
from strings, this maintains core words that are then used to identify and link duplicates.
This is a choice that considers errors should lie on the side of false
negatives rather than false positives.
}
\examples{
\dontrun{
IEADB$line <- code_linkage(IEADB$Title)
}
}
