% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_agreements.R
\name{code_type}
\alias{code_type}
\title{Code Agreement Type}
\usage{
code_type(title)
}
\arguments{
\item{title}{A character vector of treaty title}
}
\value{
A character vector of the treaty type
}
\description{
Identify the type of international agreement from titles.
Agreements can be, for example, multilateral treaties or coventions (A),
protocols (P) or amendments (E), if they contain words in title.
}
\examples{
IEADB <- dplyr::slice_sample(qEnviron::agreements$IEADB, n = 10)
IEADB$type <- code_type(IEADB$Title)
}
