% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_agreements.R
\name{code_type}
\alias{code_type}
\title{Code Agreement Type}
\usage{
code_type(x)
}
\arguments{
\item{x}{A character vector of treaty title}
}
\value{
A character vector of the type of treaty
}
\description{
Identify the type of the international agreement.
}
\examples{
IEADB$Type <- code_type(IEADB$Title)
}
