% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\alias{collapse_select}
\alias{collapse_full}
\alias{collapse_consensus}
\title{Collapse a database into a single data frame/tibble}
\usage{
collapse_select(dbase, dset)

collapse_full(dbase, key)

collapse_consensus(dbase, key)
}
\arguments{
\item{dbase}{A qPackage database object}

\item{dset}{A dataset label from within that database}

\item{key}{An ID column to collapse by}
}
\description{
Collapse a database into a single data frame/tibble
}
