% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_categories}
\alias{compare_categories}
\title{Compare categories in 'many' databases}
\usage{
compare_categories(
  database,
  dataset = NULL,
  key = "manyID",
  variable = "all",
  category = "all"
)
}
\arguments{
\item{database}{A database from one of the many packages.}

\item{dataset}{A dataset in a database from one of the many packages.
NULL by default.
That is, all datasets in the database are used.
A list of 2 or more datasets present in the database.}

\item{key}{A variable key to join datasets.
'manyID' by default.}

\item{variable}{Would you like to focus on one, or more, specific variables
present in one or more datasets in the 'many' database?
By default "all".
For multiple variables, please declare variable names as a vector.}

\item{category}{Would you like to focus on one specific code category?
By default "all" are returned.
Other options include "confirmed", "unique", "missing", "conflict",
or "majority".
For multiple variables, please declare categories as a vector.}
}
\description{
Compare categories in 'many' databases
}
\details{
Confirmed values are the same in all datasets in database.
Unique values appear once in datasets in database.
Missing values are missing in all datasets in database.
Conflict values are different in the same number of datasets in database.
Majority values have the same value in multiple, but not all,
datasets in database.
}
\examples{
compare_categories(emperors, key = "ID")
compare_categories(database = emperors, dataset = c("wikipedia", "UNRV"),
key = "ID", variable = c("Beg", "End"), category = c("conflict", "unique"))
plot(compare_categories(emperors, key = "ID"))
plot(compare_categories(database = emperors, dataset = c("wikipedia", "UNRV"),
key = "ID", variable = c("Beg", "End"), category = c("conflict", "unique")))
}
\seealso{
Other compare_: 
\code{\link{compare_data}()},
\code{\link{compare_missing}()},
\code{\link{compare_overlap}()}
}
\concept{compare_}
