% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_overlap}
\alias{compare_overlap}
\title{Compare the overlap between datasets in 'many' databases}
\usage{
compare_overlap(database, dataset = NULL, key = "manyID")
}
\arguments{
\item{database}{A database from one of the many packages.}

\item{dataset}{A dataset in a database from one of the many packages.
NULL by default.
That is, all datasets in the database are used.
A list of 2 or more datasets present in the database.}

\item{key}{A variable key to join datasets.
'manyID' by default.}
}
\value{
\code{compare_overlap()} returns a tibble with information about each dataset
and the number of overlapping observations.
}
\description{
Compare the overlap between datasets in 'many' databases
}
\details{
\code{compare_overlap()} compares the overlap between "key" observations
in each dataset in a 'many' database.
}
\examples{
\donttest{
compare_overlap(emperors, key = "ID")
plot(compare_overlap(emperors, key = "ID"))
}
}
\seealso{
Other compare_: 
\code{\link{compare_categories}()},
\code{\link{compare_data}()},
\code{\link{compare_missing}()}
}
\concept{compare_}
