% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate.R
\name{consolidate}
\alias{consolidate}
\title{Consolidate datacube into a single dataset}
\usage{
consolidate(
  datacube,
  join = c("full", "inner", "left"),
  resolve = "coalesce",
  key = NULL
)
}
\arguments{
\item{datacube}{A datacube from one of the many packages}

\item{join}{Which join procedure to use.
By default "full" so that all observations are retained,
but other options include "left" for basing the consolidated dataset
on observations present in the first dataset
(reorder the datasets to favour another dataset),
and "inner" for a consolidated dataset that includes only observations
that are present in all datasets.}

\item{resolve}{How should conflicts between observations be resolved?
By default "coalesce",
but other options include: "min", "max", "mean", "median", and "random".
"coalesce" takes the first non-NA value.
"max" takes the largest value.
"min" takes the smallest value.
"mean" takes the average value.
"median" takes the median value.
"random" takes a random value.
For different variables to be resolved differently,
you can specify the variables' names alongside
how each is to be resolved in a list
(e.g. \code{resolve = c(var1 = "min", var2 = "max")}).
In this case, only the variables named will be resolved and returned.}

\item{key}{An ID column to collapse by.
By default "manyID".
Users can also specify multiple key variables in a list.
For multiple key variables, the key variables must be present in
all the datasets in the datacube (e.g. \code{key = c("key1", "key2")}).
For equivalent key columns with different names across datasets,
matching is possible if keys are declared (e.g. \code{key = c("key1" = "key2")}).
Missing observations in the key variable are removed.}
}
\value{
A single tibble/data frame.
}
\description{
This function consolidates a set of datasets in a 'many*' package datacube
into a single dataset with some combination of the rows, columns,
and observations of the datasets in the datacube.
}
\details{
The function includes separate arguments for the rows and columns,
as well as for how to resolve conflicts for observations across datasets.
This provides users with considerable flexibility in how they combine data.
For example, users may wish to stick to units that appear in
every dataset but include variables coded in any dataset,
or units that appear in any dataset
but only those variables that appear in every dataset.
Even then there may be conflicts, as the actual unit-variable
observations may differ from dataset to dataset.
We offer a number of resolve methods that enable
users to choose how conflicts between observations are resolved.

Text variables are dropped for more efficient consolidation.
}
\examples{
\donttest{
consolidate(emperors, join = "full", resolve = "coalesce", key = "ID")
consolidate(emperors, join = "inner", resolve = "min", key = "ID")
consolidate(emperors, join = "left", resolve = "max", key = "ID")
}
}
