% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling_data.R
\name{consolidate}
\alias{consolidate}
\title{Get first non-missing}
\source{
https://stackoverflow.com/questions/40515180/dplyr-how-to-find-the-first-non-missing-string-by-groups
}
\usage{
consolidate(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A single value
}
\description{
For use with dplyr::summarise, for example
}
\details{
This function operates similarly to coalesce for columns,
that is picking the first non-missing observation,
but on observations rather than variables.
}
\examples{
dplyr::summarise(mtcars, consolidate(mtcars))
consolidate(mtcars$wt)
}
