% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate.R
\name{consolidate}
\alias{consolidate}
\title{Consolidate database into a single dataset}
\usage{
consolidate(
  database,
  rows = c("any", "every"),
  cols = c("any", "every"),
  resolve = "coalesce",
  key = "qID"
)
}
\arguments{
\item{database}{A qPackage database object}

\item{rows}{Which rows or units to retain.
By default "any" (or all) units are retained,
but another option is "every",
which retains only those units that appear in all parent datasets.}

\item{cols}{Which columns or variables to retain.
By default "any" (or all) variables are retained,
but another option is "every",
which retains only those variables that appear in all parent datasets.}

\item{resolve}{How conflicts between observations should be resolved.
Currently only "coalesce" is offered, which takes the first non-NA value,
but options should soon include
"max", "min", "mean", "mode" and "median", as well as "append".}

\item{key}{An ID column to collapse by. By default "qID".}
}
\value{
A single tibble/data frame.
}
\description{
This function collapses a set or database of (q)datasets
into a single dataset with some combination of the
rows, columns, and observations of the parent datasets.
The function includes separate arguments for the rows and columns,
as well as for how to resolve conflicts in observations across datasets.
This provides users with considerable flexibility in how they combine qData.
For example, users may wish to stick to units that appear in every dataset
but include variables coded in any dataset,
or units that appear in any dataset
but only those variables that appear in every dataset.
Even then there may be conflicts, as the actual unit-variable observations
may differ from dataset to dataset.
Here we (will) offer a number of resolve methods that enable users to choose
how conflicts between observations are resolved.
}
\examples{
consolidate(agreements, "any", "any")
consolidate(states, "every", "every", key = "ID")
}
