% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_data.R
\name{data_contrast}
\alias{data_contrast}
\title{Contrasts data structures in qPackages}
\usage{
data_contrast(pkg, database = NULL, dataset = NULL)
}
\arguments{
\item{pkg}{character string of the qPackage to report data on. Required
input.}

\item{database}{vector of character strings of the qPackage to report data a
specific database in a qPackage. If NULL, report_data returns a summary
of all databases in the qPackage. NULL by default.}

\item{dataset}{character string of the qPackage to report data on a specific
dataset in a specific database of a qPackage. If NULL and database is
specified, returns database level metadata. NULL by default.}
}
\value{
A dataframe with the data report
}
\description{
Allows users to quickly compare processed datasets before analysing them.
}
\details{
This function displays information about databases and datasets
contained in them. Namely the number of unique ID's, the percentage of
missing data, the number of observations, the number of variables, the
minimum beginning date and the maximum ending date as well as the most direct
URL to the original dataset.
}
\examples{
data_contrast(pkg = "qStates", database = "states", dataset = "COW")
}
