% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_plot.R
\name{db_plot}
\alias{db_plot}
\title{Plot database profile}
\usage{
db_plot(database, key)
}
\arguments{
\item{database}{A many database.}

\item{key}{A joining key.}
}
\value{
A plot with a profile of the database.
}
\description{
Plot database profile
}
\examples{
db_plot(emperors, "ID")
}
