% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{depends}
\alias{depends}
\title{Helper function for loading and, if necessary, installing CRAN packages}
\usage{
depends(packages)
}
\arguments{
\item{packages}{Character vector of packages to install from CRAN}
}
\value{
Loads and, if necessary, first installs CRAN packages
}
\description{
Helper function for loading and, if necessary, installing CRAN packages
}
