% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qData-emperors.R
\docType{data}
\name{emperors}
\alias{emperors}
\title{emperors database documentation}
\format{
The emperors database is a list that contains the
following 1 datasets: wikipedia.
For more information and references to each of the datasets used,
please use the \code{data_source()}, \code{data_contrast()}, and \code{data_evolution()}
functions.
\describe{
\item{wikipedia: }{A dataset with 68 observations and the following 16 variables: index, name.full, birth, death, birth.cty, birth.prv, rise, cause, killer, dynasty, era, notes, verif.who, ID, Beg, End.}
}
}
\usage{
emperors
}
\description{
emperors database documentation
}
\keyword{datasets}
