% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emperors.R, R/manydata-emperors.R
\docType{data}
\name{emperors}
\alias{emperors}
\title{Emperors database documentation}
\format{
The emperors database is a list that contains the
following 3 datasets: wikipedia, UNRV, britannica.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()}.
functions.
\describe{
\item{wikipedia: }{A dataset with 68 observations and the
following 15 variables:
ID, Beg, End, FullName, Birth, Death,
CityBirth, ProvinceBirth, Rise, Cause,
Killer, Dynasty, Era, Notes, Verif.}
\item{UNRV: }{A dataset with 99 observations and the
following 7 variables: ID, Beg, End,
Birth, Death, FullName, Dynasty.}
\item{britannica: }{A dataset with 87 observations and the
following 3 variables: ID, Beg, End.}
}

The emperors database is a list that contains the
following 3 datasets: wikipedia, UNRV, britannica.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{wikipedia: }{A dataset with 68 observations and the following
16 variables: index, name, name.full, birth, death, birth.cty, birth.prv,
rise, reign.start, reign.end, cause, killer, dynasty, era, notes, verif.who.}
\item{UNRV: }{A dataset with 99 observations and the following
7 variables: ID, Beg, End, Birth, Death, FullName, Dynasty.}
\item{britannica: }{A dataset with 87 observations and the following
3 variables: ID, Beg, End.}
}
}
\usage{
emperors

emperors
}
\description{
Emperors database documentation

emperors database documentation
}
\details{
\preformatted{## $wikipedia
## ── Data Summary ────────────────────────
##                            Values
## Name                       X[[i]]
## Number of rows             68    
## Number of columns          15    
## _______________________          
## Column type frequency:           
##   character                13    
##   mdate                    2     
## ________________________         
## Group variables            None  
## 
## ── Variable type: character ────────────────────────────────────────────────────
##    skim_variable n_missing complete_rate   min   max empty n_unique whitespace
##  1 ID                    0         1         4    21     0       68          0
##  2 FullName              0         1        26    72     0       67          0
##  3 Birth                 5         0.926     4    22     0       60          0
##  4 Death                 0         1        10    11     0       66          0
##  5 CityBirth            17         0.75      4    16     0       30          0
##  6 ProvinceBirth         0         1         5    18     0       18          0
##  7 Rise                  0         1         8    31     0        8          0
##  8 Cause                 0         1         7    14     0        7          0
##  9 Killer                0         1         4    16     0       14          0
## 10 Dynasty               0         1         7    14     0        8          0
## 11 Era                   0         1         8    10     0        2          0
## 12 Notes                22         0.676    22   104     0       23          0
## 13 Verif                57         0.162    22    23     0        2          0
## 
## ── Variable type: mdate ────────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate n_empty n_unique max        min       
## 1 Beg                   0             1       0       58 0379-01-01 -026-01-16
## 2 End                   0             1       0       65 0395-01-17 0014-08-19
## 
## $UNRV
## ── Data Summary ────────────────────────
##                            Values
## Name                       X[[i]]
## Number of rows             99    
## Number of columns          7     
## _______________________          
## Column type frequency:           
##   character                5     
##   mdate                    2     
## ________________________         
## Group variables            None  
## 
## ── Variable type: character ────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate   min   max empty n_unique whitespace
## 1 ID                    0         1         4    28     0       98          0
## 2 Birth                 0         1         1     5     0       70          0
## 3 Death                 0         1         1     4     0       79          0
## 4 FullName              5         0.949     1   141     0       91          0
## 5 Dynasty              37         0.626     1    98     0       28          0
## 
## ── Variable type: mdate ────────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate n_empty n_unique max        min       
## 1 Beg                   0             1       0       67 0491-12-31 -027-01-01
## 2 End                   0             1       0       78 0518-12-31 -014-01-01
## 
## $britannica
## ── Data Summary ────────────────────────
##                            Values
## Name                       X[[i]]
## Number of rows             87    
## Number of columns          3     
## _______________________          
## Column type frequency:           
##   character                1     
##   mdate                    2     
## ________________________         
## Group variables            None  
## 
## ── Variable type: character ────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate   min   max empty n_unique whitespace
## 1 ID                    0             1     4    30     0       87          0
## 
## ── Variable type: mdate ────────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate n_empty n_unique max        min       
## 1 Beg                   0             1       0       69 0475-12-31 0014-01-01
## 2 End                   0             1       0       76 0491-12-31 0037-01-01
}
}
\keyword{datasets}
