% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_bilaterals.R
\name{extraction}
\alias{extraction}
\alias{extract_bilaterals}
\alias{extract_multilaterals}
\title{Extracting adjacency edgelist from bilaterals in a mixed edgelist}
\usage{
extract_bilaterals(membs)

extract_multilaterals(membs)
}
\description{
Extracting adjacency edgelist from bilaterals in a mixed edgelist
}
\examples{
\dontrun{
membs <- qEnviron::memberships$IEADB_MEM
extract_bilaterals(membs)
extract_multilaterals(membs)
}
}
