% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate.R
\name{favour}
\alias{favour}
\alias{favor}
\title{Favour datasets in a database}
\usage{
favour(database, dataset)

favor(database, dataset)
}
\arguments{
\item{database}{A many database}

\item{dataset}{The name of one, or more, datasets within the database
to be favoured over others.}
}
\value{
The database with datasets re-ordered accordingly
}
\description{
Favour datasets in a database
}
\details{
The dataset declared becomes the reference for
the first non NA value.
If more than one dataset is declared,
please add datasets increasing order of importance
(.i.e. last dataset should be favoured over previous).
}
\examples{
favour(emperors, "UNRV")
favour(emperors, c("wikipedia", "UNRV", "britannica"))
}
