% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate.R
\name{favour}
\alias{favour}
\alias{favor}
\title{Favour datasets in a datacube}
\usage{
favour(datacube, dataset)

favor(datacube, dataset)
}
\arguments{
\item{datacube}{A many datacube}

\item{dataset}{The name of one, or more, datasets within the datacube
to be favoured over others.}
}
\value{
The datacube with datasets re-ordered accordingly
}
\description{
Favour datasets in a datacube
}
\details{
The dataset declared becomes the reference for
the first non NA value.
If more than one dataset is declared,
please list datasets in increasing order of importance
(.i.e. last dataset should be favoured over previous).
}
\examples{
favour(emperors, "UNRV")
favour(emperors, c("wikipedia", "UNRV", "britannica"))
}
