% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_get.R
\name{get_packages}
\alias{get_packages}
\title{Find and download packages in the qData ecosystem}
\usage{
get_packages(pkg)
}
\arguments{
\item{pkg}{A character vector of package names}
}
\value{
If no package name is provided, this function prints a table (tibble) to the console
with details on packages that are currently available within the qData ecosystem.
This includes the name and description of the package,
the latest installed and release version number, and the latest release date,
and a string of contributors.

If one or more package names are provided, these will be installed from Github.
}
\description{
Find and download packages in the qData ecosystem
}
\details{
The function finds and download other packages that belong to the qData ecosystem
of data packages. It allows for users to rapidly access the names and other descriptive information
of these packages by simply calling the function. If users intend to download a package from
the ecosystem, they can to type the package name within the function.
}
\examples{
\dontrun{
get_packages() # This prints a table (tibble) 
# to the console with details on currently available qPackages
get_packages("qStates") # This downloads and installs from Github 
# the latest version of one or more qPackages
}
}
