% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{get_packages}
\alias{get_packages}
\title{Find and download packages in the qDatr ecosystem}
\usage{
get_packages(pkg)
}
\arguments{
\item{pkg}{A character vector of package names}
}
\value{
If no package name is provided, this function prints a table (tibble) to the console
with details on packages that are currently available within the qDatr ecosystem.
If one or more package names are provided, these will be installed from
}
\description{
Find and download packages in the qDatr ecosystem
}
\details{
The function finds and download other packages that belong to the qDatr ecosystem
of data packages. It allows for users to rapidly access the names and other descriptive information
of these packages by simply calling the function. If users intend to download a package from the ecosystem,
they can to type the package name within the function.
}
\examples{
\dontrun{
get_packages() # This prints a table (tibble) to the console with details on the current
get_packages("qStates") # This downloads and installs from github the latest version of a qPackage
}
}
