% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_get.R
\name{get_packages}
\alias{get_packages}
\title{Find and download packages in the many packages universe}
\usage{
get_packages(pkg)
}
\arguments{
\item{pkg}{A character vector of package names or number of a package}
}
\value{
If no package name is provided,
this function prints a table (tibble)
to the console with details on packages
that are currently available within
the many universe.
This includes the name and description of the package,
the latest installed and release version number,
and the latest release date.
It also include a list of numbers which orders
the package and can be used to load the respective
package instead of the name.
If one or more package names are provided,
these will be installed from Github.
}
\description{
Find and download packages in the many packages universe
}
\details{
The function finds and download other packages
that belong to the many universe of packages.
It allows users to rapidly access the names and other
descriptive information of these
packages by simply calling the function.
If users intend to download and install
a package from the universe,
they can type the package name within the function.
}
