% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{import_data}
\alias{import_data}
\title{Imports and establishes preparation of raw data}
\usage{
import_data(
  dataset = NULL,
  database = NULL,
  path = NULL,
  delete_original = FALSE,
  open = rlang::is_interactive()
)
}
\arguments{
\item{dataset}{Intended (short)name of the dataset}

\item{database}{Intended (short)name of the dataset}

\item{path}{Path to raw data file. If left unspecified, a dialog box is raised to select the file via the system}

\item{delete_original}{Does not delete original files by default.}

\item{open}{Whether the resulting preparation script will be opened}
}
\value{
This function returns a data-raw folder containing the data imported as well as a script in the R directory
to guide preparation of data using qDatr.
}
\description{
Create a data-raw folder and provide templates that make easier for setting up the data cleaning
and wrangling, consistent with the qDatr ecosystem
}
\details{
The function helps importing raw data into q package while providing a template that facilitates
data cleaning and wrangling, consistent with the qDatr ecosystem. The function can be used without specifying
a path to the file. In that case an interactive dialog box will be openend and the data file can be manually selected.
The script provided to help with data cleaning and wrangling contain suggestions on how to properly load the data into
the environment.
}
\examples{
\dontrun{
<<<<<<< HEAD
qDatr::import_data("cow")
=======
qDatr::import_data(dataset = "cow", database = "states")
>>>>>>> 8b34cf68755f4473b6e807a99a9695e741fe4b7e
}
}
