% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreements_plot.R
\name{network_map}
\alias{network_map}
\title{Plot geographical networks}
\usage{
network_map(
  dataset,
  actor = "StateID",
  treaty_type = NULL,
  date = "2019-12-31",
  theme = "light"
)
}
\arguments{
\item{dataset}{A dataset from one of the many packages
or a "consolidated" database.}

\item{actor}{An actor variable.
"StateID", by default.}

\item{treaty_type}{The type of treaties to be returned.
NULL, by default.
Other options are "bilateral" or "multilateral".}

\item{date}{String date from the network snapshot.
Used by \code{{cshapes}} to plot the correct map.
By default, 2019-12-31.
Date can be between 1886-01-01 and 2019-12-31.}

\item{theme}{Theme you would like to use to plot the graph.
bey defalt, "light".
Available themes are "light", "dark", and "earth".}
}
\value{
A map of a country level geographical network.
}
\description{
Creates a plot of the a unimodal geographical network at a single point
in time.
}
\examples{
\donttest{
#memberships <- dplyr::filter(manyenviron::memberships$ECOLEX_MEM,
#Beg > "2000-01-01" & Beg < "2000-12-12")
#network_map(memberships, actor = "CountryID") +
#ggplot2::labs(title = "International Environmental Treaties Signed in the year 2000",
#subtitle = "Ecolex data")
}
}
