% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{new_author}
\alias{new_author}
\title{Helper function for adding an author to the current package}
\usage{
new_author(
  given = NULL,
  family = NULL,
  email = NULL,
  role = NULL,
  comment = NULL
)
}
\arguments{
\item{given}{Character string of the author's name}

\item{family}{Character string of the author's surname}

\item{email}{Character string of the author's email}

\item{role}{Character vector of role(s) the author has in the project}

\item{comment}{Character vector of the author's miscellaneous information
such as his/her institution or his/her ORCID number.}
}
\value{
Adds a new author to the description file of the package
}
\description{
Helper function for adding an author to the current package
}
\details{
The function adds an author to the description file of the current
package.
}
