% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreements_plot.R
\name{plot_agreements}
\alias{plot_agreements}
\alias{agreements_plot}
\alias{membership_plot}
\alias{lineage_plot}
\alias{map_plot}
\title{Plot agreements network}
\usage{
agreements_plot(dataset, treaty_type = NULL, layout = "circle")

membership_plot(
  dataset,
  actor = "stateID",
  treaty_type = NULL,
  layout = "circle"
)

lineage_plot(dataset, treaty_type = NULL)

map_plot(
  dataset,
  actor = "StateID",
  treaty_type = NULL,
  date = "2019-12-31",
  theme = "light"
)
}
\arguments{
\item{dataset}{A dataset from one of the many packages
or a "consolidated" database.}

\item{treaty_type}{The type of treaties to be returned.
NULL, by default.
Other options are "bilateral" or "multilateral".}

\item{layout}{How do you want the plot to look like?
An igraph layout algorithm, currently defaults to 'concentric'.
Some other options are 'stress', 'bipartite', and 'alluvial'.
For more information please check ´?manynet::autographr´.}

\item{actor}{An actor variable.
"StateID", by default.}

\item{date}{String date from the network snapshot.
Used by \code{{cshapes}} to plot the correct map.
By default, 2019-12-31.
Date can be between 1886-01-01 and 2019-12-31.}

\item{theme}{Theme you would like to use to plot the graph.
bey defalt, "light".
Available themes are "light", "dark", and "earth".}
}
\value{
A network of agreements' relations.

A network of agreements' memberships.

A plot of agreements' lineages.

A map of a country level geographical network.
}
\description{
Facilitates plotting of 'many' data.
}
\details{
Creates a plot of the a unimodal geographical network at a
single point in time.
}
\examples{
\donttest{
#agreements <- dplyr::filter(manyenviron::agreements$ECOLEX,
#Beg > "2000-01-01" & Beg < "2000-12-12")
#agreements_plot(agreements)
}
\donttest{
#memberships <- dplyr::filter(manyenviron::memberships$ECOLEX_MEM,
#Beg > "2000-01-01" & Beg < "2000-06-12")
#membership_plot(memberships)
}
\donttest{
#dataset <- dplyr::filter(manyenviron::agreements$HUGGO, Beg > "2000-01-01",
#Beg < "2010-12-31")
#lineage_plot(dataset)
}
\donttest{
#memberships <- dplyr::filter(manyenviron::memberships$ECOLEX_MEM,
#Beg > "2000-01-01" & Beg < "2000-12-12")
#map_plot(memberships, actor = "stateID") +
#ggplot2::labs(title = "Bilateral International Environmental Treaties Signed in the year 2000",
#subtitle = "Ecolex data")
}
}
