% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate.R
\name{pluck}
\alias{pluck}
\title{Selects a single dataset from a datacube}
\usage{
pluck(.x, ..., .default = NULL)
}
\value{
The selected dataset
}
\description{
This function is reexported from the \code{{purrr}} package.
It allows users to select a single dataset from one
of the datacubes available across the 'many* packages'.
.  It additionally invites users to cite the
}
\examples{
\donttest{
pluck(emperors, "UNRV")
}
}
