% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{qtemplate}
\alias{qtemplate}
\title{Helper function for finding and rendering templates}
\usage{
qtemplate(
  template,
  save_as = path,
  data = list(),
  ignore = FALSE,
  open = FALSE,
  package = "qDatr"
)
}
\arguments{
\item{template}{Template called}

\item{save_as}{Path to where the rendered template should be saved}

\item{data}{Any elements to be entered into the template via Whisker}

\item{ignore}{For use with usethis::use_build_ignore()}

\item{open}{Whether the resulting template will be opened}

\item{package}{Package called}
}
\value{
A rendered template, saved into the correct folder
}
\description{
Helper function for finding and rendering templates from the qDatr package
}
\details{
This function is an adaptation of the usethis variant
for use in the qDatr ecosystem.
}
\examples{
\dontrun{
TODO
}
}
