% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{recollect}
\alias{recollect}
\title{Pastes unique string vectors}
\usage{
recollect(x, collapse = "_")
}
\arguments{
\item{x}{A vector}

\item{collapse}{String indicating how elements separated}
}
\value{
A single value
}
\description{
For use with dplyr::summarise, for example
}
\details{
This function operates similarly to reunite,
but instead of operating on columns/observations,
it pastes together unique rows/observations.
}
\examples{
data <- data.frame(ID = c(1,2,3,3,2,1))
data1 <- data.frame(ID = c(1,2,3,3,2,1), One = c(1,NA,3,NA,2,NA))
recollect(data$ID)
recollect(data1$One)
}
