% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate.R, R/zzz.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{pluck}
\alias{\%>\%}
\title{Selects a single dataset from a datacube}
\value{
The selected dataset
}
\details{
This function is reexported from the purrr package.
It allows users to select a single dataset from one
of the datacubes available across the 'many* packages'.
}
\examples{
\donttest{
pluck(emperors, "UNRV")
}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}}

  \item{purrr}{\code{\link[purrr]{pluck}}}
}}

