% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_titles.R
\name{repaint}
\alias{repaint}
\title{Fills missing data by lookup}
\usage{
repaint(df, id, var)
}
\arguments{
\item{df}{a dataframe}

\item{id}{a string identifying a column in the dataframe for indexing}

\item{var}{a string identifying a column or columns in the dataframe
to be filled}
}
\value{
a dataframe
}
\description{
Fills missing data where known by other observations with the same id/index
}
\examples{
\dontrun{
myData <- repaint(myData, id="StatID", var=c("Area","Region")
}
}
