% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_data.R
\name{report}
\alias{report}
\alias{data_source}
\alias{data_contrast}
\title{Set of data structure exploration functions for users}
\usage{
data_source(pkg, database = NULL, dataset = NULL)

data_contrast(pkg, database = NULL, dataset = NULL)
}
\arguments{
\item{pkg}{character string of the qPackage to report data on.
Required input.}

\item{database}{vector of character strings of the qPackage to
report data a specific database in a qPackage.
If NULL, the function returns a summary of all databases in the qPackage.
NULL by default for \code{data_source()} and \code{data_contrast()}. Required for
\code{data_evolution()}.}

\item{dataset}{character string of the qPackage to report data on a specific
dataset in a specific database of a qPackage.
If NULL and database is specified, returns database level metadata.
NULL by default for \code{data_source()} and \code{data_contrast()}. Required for
\code{data_evolution()}.}
}
\value{
A dataframe with the data sources

A list with the desired metadata to compare various datasets in the
qVerse.
}
\description{
The report family of functions allows users
to quickly get information about and compare several
aspects of a q Packages, databases and datasets.
}
\details{
\code{data_source()} displays names of the database/datasets and
source material of qPackage data.

\code{data_contrast()} displays information about databases
and datasets contained in them.
Namely the number of unique ID's, the percentage of
missing data, the number of observations, the number of variables,
the minimum beginning date and the maximum ending date as well as
the most direct URL to the original dataset.
}
\examples{
data_source(pkg = "qData", database = "states", dataset = "COW")
data_contrast(pkg = "qData", database = "states")
}
