% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_data.R
\name{report_data}
\alias{report_data}
\title{To quickly report on the databases and datasets within a qPackage}
\usage{
report_data(pkg)
}
\arguments{
\item{pkg}{character string of the qPackage to report data on}
}
\description{
To quickly report on the databases and datasets within a qPackage
}
