% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_data.R
\name{report_data}
\alias{report_data}
\title{Reports on qPackage data}
\usage{
report_data(pkg, database = NULL, dataset = NULL)
}
\arguments{
\item{pkg}{character string of the qPackage to report data on. Mandatory input.}

\item{database}{character string of the qPackage to report data a specific database in a
qPackage. If Null, report_data returns a summary of all databases in the qPackage.
Null by default.}

\item{dataset}{character string of the qPackage to report data on a specific dataset
in a specific database of a qPackage. If Null and database is specified, returns database
level metadata. Null by default.}
}
\value{
A dataframe with the data report
}
\description{
To quickly report metadata on the databases and datasets within a specific qPackage
}
\examples{
report_data(pkg = "qStates", database = "states", dataset = "COW")
}
