% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{resolve}
\alias{resolve}
\alias{resolve_min}
\alias{resolve_max}
\alias{resolve_mean}
\alias{resolve_median}
\alias{resolve_mode}
\alias{resolve_dates}
\title{Resolve ambiguous variables according to preferred output}
\usage{
resolve_min(var)

resolve_max(var)

resolve_mean(var)

resolve_median(var)

resolve_mode(var)

resolve_dates(var, resolve = NULL)
}
\arguments{
\item{var}{Ranged dates variable returned by \code{standardise_dates()}}

\item{resolve}{How do you want date ranges to be solved?}
}
\value{
a date column
}
\description{
The resolve family of function resolves various ambiguous data formats
according to user preferences.For date ranges the function resolves internal ranges
according to preferred output before moving into resolving across datasets in a database.
}
\details{
\code{resolve_dates()} resolves ranged dates into single vectors
This function resolves ranged dates created with \code{standardise_dates()}
by the choice resolve of minimum, maximun or mean dates.
}
\examples{
dates <- data.frame(dates = c("2010-01-01:2010-12-31", "1816-01-01:1916-01-01"))
resolve_dates(dates$dates, resolve = "min")
resolve_dates(dates$dates, resolve = "max")
resolve_dates(dates$dates, resolve = "mean")
}
