% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{resolving}
\alias{resolving}
\alias{resolve_coalesce}
\alias{resolve_unite}
\alias{resolve_min}
\alias{resolve_max}
\alias{resolve_random}
\alias{resolve_precision}
\alias{resolve_mean}
\alias{resolve_mode}
\alias{resolve_median}
\alias{resolve_consensus}
\title{Resolving multiple observations of the same variable into one}
\usage{
resolve_coalesce(.data, vars)

resolve_unite(.data, vars, na.rm = TRUE)

resolve_min(.data, vars, na.rm = TRUE)

resolve_max(.data, vars, na.rm = TRUE)

resolve_random(.data, vars, na.rm = TRUE)

resolve_precision(.data, vars)

resolve_mean(.data, vars, na.rm = TRUE)

resolve_mode(.data, vars, na.rm = TRUE)

resolve_median(.data, vars, na.rm = TRUE)

resolve_consensus(.data, vars, na.rm = TRUE)
}
\arguments{
\item{.data}{A data frame or tibble containing the variables.}

\item{vars}{A vector of variables from \code{.data} to be resolved or converged.}

\item{na.rm}{Logical whether missing values (NAs) should be removed
before operation of the function.
Note that unlike how the \code{na.rm} argument operates in functions in
base R, e.g. \code{max()}, here the default is TRUE.}
}
\description{
This family of functions provides row-wise summarization for
data frames or tibbles,
returning a single value per row based on specified columns.
They are useful for tasks like extracting typical or summary values from
multiple variables, simplifying wide data structures,
and imputing representative values.
}
\section{Unite}{

Uniting returns all the unique values as a set,
separated by commas and contained within braces.
Note that uniting always returns a character/string vector,
which enables it to accommodate different classes of variables.
The order of the values reflects their first appearance;
that is, they are not ordered by increasing value.
}

\examples{
test <- data.frame(bloop.x = c(1,6,NA), 
                   bloop.y = c(2,NA,3), 
                   bloop = c(NA,3.1,4.1))
test
resolve_coalesce(test)
resolve_unite(test)
resolve_min(test)
resolve_max(test)
resolve_random(test)
resolve_precision(test)
resolve_mean(test)
test2 <- cbind(test, bloopy = c(2,NA,3))
resolve_mode(test2)
resolve_median(test2)
resolve_consensus(test[,1:2])
}
