% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{resolving}
\alias{resolving}
\alias{resolving_coalesce}
\alias{resolving_unite}
\title{Resolving multiple observations of the same variable into one}
\usage{
resolving_coalesce(.data, vars)

resolving_unite(.data, vars)

}
\description{
Resolving multiple observations of the same variable into one
}
\section{Unite}{

Note that uniting always returns a character/string vector.
Values are separated by commas and a set is contained within braces.
}

\examples{
test <- data.frame(bloop.x = c(1,6,NA), 
                   bloop.y = c(2,NA,3), bloop = c(NA,3,4))
resolving_coalesce(test)
resolving_unite(test)
}
