% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_posterior_samples.R
\name{run_MH_MCMC}
\alias{run_MH_MCMC}
\title{Simulate posterior samples using MCMC}
\usage{
run_MH_MCMC(startval, iterations, dat_obs, dat_exp, eta)
}
\arguments{
\item{startval}{Initial values..}

\item{iterations}{Number of MCMC iterations..}

\item{dat_obs}{Observational dataset.}

\item{dat_exp}{Experimental dataset.}

\item{eta}{Learning rate on the observational data..}
}
\value{
A table of posterior samples.
}
\description{
Simulate posterior samples using MCMC
}
\examples{
theta_sim_eta05 <- run_MH_MCMC(c(1,0.5), 1000, dat_obs = dat_obs, dat_exp = dat_exp, eta = 0.5)
}
