% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_setup.R
\name{setup_package}
\alias{setup_package}
\title{Create a new package in the qData ecosystem}
\usage{
setup_package(
  package = NULL,
  name = NULL,
  role = NULL,
  orcid = NULL,
  update = TRUE,
  path = getwd()
)
}
\arguments{
\item{package}{A string giving the desired name of the package,
must start with "q"}

\item{name}{A list of vectors giving the package
author(s)' name(s). Authors(s)last name(s) and first
name(s) are separated by a comma.}

\item{role}{A list of vectors of the roles the package authors have
in the project. If there are no roles declared,
roles are set contributor.}

\item{orcid}{A vector of strings of all the ORCID numbers of the authors.
Needs \code{{rorcid}} package to be installed. Takes precedence over manual
entries if specified.}

\item{update}{A logical indicating whether existing files should be
overwritten, by default TRUE.}

\item{path}{A string, if missing default is path to the working directory}
}
\value{
A new package structure
}
\description{
Creates a new package in, and consistent with, the qData ecosystem
}
\details{
The function establishes many of the required files and
folder structures required for a qData-consistent data package.
}
\examples{
\dontrun{
setup_package("qStates", name = "Hollway, James"))
}
\dontrun{
setup_package("qStates", 
orcid = c("0000-0002-8361-9647"))
}
}
