% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_titles.R
\name{standardise_titles}
\alias{standardise_titles}
\title{Capitalises all words}
\usage{
standardise_titles(s, strict = FALSE)
}
\arguments{
\item{s}{A string}

\item{strict}{By default FALSE}
}
\value{
A capitalised, trimmed string
}
\description{
Capitalises all words in a string to enable comparison
}
\details{
The function capitalises all words in the strings passed to it,
as well as trimming all white space from the starts and ends of the strings.
}
\examples{
\dontrun{
e <- standardise_titles("A treaty concerning things")
e==c("A Treaty Concerning Things")
}
}
