% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states_database.R
\docType{data}
\name{states}
\alias{states}
\title{states Database Documentation}
\format{
The states database is a list that contains a shorter
version of the following 3 datasets: GW, ISD, COW.
}
\usage{
states
}
\description{
states Database Documentation
}
\details{
These datasets are to be used for testing and
running examples and vignettes in qData.
For the complete database and more datasets on states,
please refer to the \code{{qStates}} package.
\describe{
\item{GW: }{A dataset with 150 observations and the following 5 variables:
COW_Nr, ID, Beg, End, Label.}
\item{ISD: }{A dataset with 150 observations and the following 7 variables:
COW_Nr, ID, Beg, End, Label, Micro, New.State.}
\item{COW: }{A dataset with 150 observations and the following 5 variables:
COW_Nr, ID, Beg, End, Label.}
}
}
\keyword{datasets}
