% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{use_qData}
\alias{use_qData}
\title{Save a cleaned data object in the new q package}
\usage{
use_qData(...)
}
\arguments{
\item{...}{Unquoted names of existing objects to save}
}
\description{
Save a cleaned data object, consistent with the qDatr ecosystem, ready to be lazy-loaded
and create scripts for documenting and testing that object within the new q package
}
\examples{
\dontrun{
TODO
}
}
