% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{use_qData_raw}
\alias{use_qData_raw}
\title{Createa a dataraw file the new q package}
\usage{
use_qData_raw(name = "DATASET", open = rlang::is_interactive())
}
\arguments{
\item{name}{name of dataset saved}

\item{open}{}
}
\value{
A dataraw folder
}
\description{
Creates a dataraw file and provides templates that make it consistent with the qDatr ecosystem
}
\details{
The function loads raw data into a q package
}
\examples{
\dontrun{
TODO
}
}
